services/admin/requirements.txt
confluent-kafka==2.3.0
python-decouple==3.8
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
requests==2.31.0
